requirements.txt
flask
python-dotenv
requests
plaid-python
